Mild refactoring for clarity
