--- src/vs/workbench/contrib/positronConsole/browser/positronConsoleFindCommandIds.ts
@@ -1,14 +1,0 @@
-/*---------------------------------------------------------------------------------------------
- *  Copyright (C) 2026 Posit Software, PBC. All rights reserved.
- *  Licensed under the Elastic License 2.0. See LICENSE.txt for license information.
- *--------------------------------------------------------------------------------------------*/
-
-export const enum PositronConsoleFindCommandId {
-	FindFocus = 'workbench.action.positronConsole.findFocus',
-	FindHide = 'workbench.action.positronConsole.findHide',
-	FindNext = 'workbench.action.positronConsole.findNext',
-	FindPrevious = 'workbench.action.positronConsole.findPrevious',
-	ToggleFindRegex = 'workbench.action.positronConsole.toggleFindRegex',
-	ToggleFindWholeWord = 'workbench.action.positronConsole.toggleFindWholeWord',
-	ToggleFindCaseSensitive = 'workbench.action.positronConsole.toggleFindCaseSensitive',
-}
